requirements.txt
# Upgrade OpenAI SDK so attachments + responses work
openai>=1.51.0

streamlit>=1.37.0
pydantic>=2.7.0
python-dotenv>=1.0.1
orjson>=3.10.7
httpx>=0.27.0
rich>=13.7.1
markdownify>=0.13.1
duckduckgo-search>=6.2.12

# Azure AI Foundry Agent auth (Managed Identity / Entra ID)
azure-identity>=1.16.0
